airflow/requirements.txt
apache-airflow==2.10.5
pymongo>=4.6.0
psycopg2-binary>=2.9.9
sqlalchemy>=1.4.36,<2.0
python-dotenv>=1.0.0
requests>=2.31.0
